// backoffice-frontend/src/app/models/product.model.ts
export class Product {
    comments: string;
    category: number;
    availability: boolean;
    price: number;
    price_on_sale: number;
    discount: number;
    sale: boolean;
    owner: string;
    unit: string;
    name: string;
    quantity_stock: number;
    quantity_sold: number;
    tigID: number;

    updated_quantity_stock: number;
    updated_discount: number;
}